refactor: Rename routes to regular-routes to avoid confusion with Gatsby routes
