+++ src/pages/regular-routes/uk-to-france.tsx
@@ -3,15 +3,15 @@ import { Helmet } from 'react-helmet'
 import { StaticImage } from 'gatsby-plugin-image'
 import RouteLayout from '../../layouts/Route'
 import logoSrc from '../../images/lettermark_blue.svg'
-import netIcon from '../../images/routes/icons/noun_net_2428552.svg'
-import mapIcon from '../../images/routes/icons/noun_Maps_3610706.svg'
-import truckIcon from '../../images/routes/icons/noun_Truck_1731459.svg'
-import heartBillIcon from '../../images/routes/icons/noun_Heart_Bill_98293.svg'
-import boxIcon from '../../images/routes/icons/noun_Box_1897159.svg'
-import sackIcon from '../../images/routes/icons/noun_sack_1468937.svg'
-import vanIcon from '../../images/routes/icons/noun_Van_485356.svg'
-import halfPalletIcon from '../../images/routes/icons/noun_Pallet_3364535.svg'
-import palletIcon from '../../images/routes/icons/noun_Pallet_3307940.svg'
+import netIcon from '../../images/regular-routes/icons/noun_net_2428552.svg'
+import mapIcon from '../../images/regular-routes/icons/noun_Maps_3610706.svg'
+import truckIcon from '../../images/regular-routes/icons/noun_Truck_1731459.svg'
+import heartBillIcon from '../../images/regular-routes/icons/noun_Heart_Bill_98293.svg'
+import boxIcon from '../../images/regular-routes/icons/noun_Box_1897159.svg'
+import sackIcon from '../../images/regular-routes/icons/noun_sack_1468937.svg'
+import vanIcon from '../../images/regular-routes/icons/noun_Van_485356.svg'
+import halfPalletIcon from '../../images/regular-routes/icons/noun_Pallet_3364535.svg'
+import palletIcon from '../../images/regular-routes/icons/noun_Pallet_3307940.svg'
 
 import palletStorageBackground from './pallet-storage.jpg'
 import forkliftLoadingBackground from './forklift-loading.jpg'
@@ -293,7 +293,7 @@ const UkToFrance: FC = () => {
                 <div className="tile-icon mx-auto">
                   <StaticImage
                     className="icon icon--responsive"
-                    src="../../images/routes/mrs-logo.256.jpg"
+                    src="../../images/regular-routes/mrs-logo.256.jpg"
                     alt="Frontline Group Logo: Mobile Refugee Support (MRS)"
                   />
                 </div>
@@ -306,7 +306,7 @@ const UkToFrance: FC = () => {
                 <div className="tile-icon mx-auto">
                   <StaticImage
                     className="icon icon--responsive rounded-full"
-                    src="../../images/routes/ca-logo.256.jpg"
+                    src="../../images/regular-routes/ca-logo.256.jpg"
                     alt="Frontline Group Logo: Collective Aid (CA)"
                   />
                 </div>
@@ -319,7 +319,7 @@ const UkToFrance: FC = () => {
                 <div className="tile-icon mx-auto">
                   <StaticImage
                     className="icon icon--responsive rounded-full"
-                    src="../../images/routes/u56-logo.256.jpg"
+                    src="../../images/regular-routes/u56-logo.256.jpg"
                     alt="Frontline Group Logo: Utopia 56 (U56)"
                   />
                 </div>
@@ -332,7 +332,7 @@ const UkToFrance: FC = () => {
                 <div className="tile-icon mx-auto">
                   <StaticImage
                     className="icon icon--responsive rounded-full"
-                    src="../../images/routes/cfc-logo.256.jpg"
+                    src="../../images/regular-routes/cfc-logo.256.jpg"
                     alt="Frontline Group Logo: Calais Food Collective (CFC)"
                   />
                 </div>
@@ -345,7 +345,7 @@ const UkToFrance: FC = () => {
                 <div className="tile-icon mx-auto">
                   <StaticImage
                     className="icon icon--responsive"
-                    src="../../images/routes/rck-logo.256.jpg"
+                    src="../../images/regular-routes/rck-logo.256.jpg"
                     alt="Frontline Group Logo: Refugee Community Kitchen (RCK)"
                   />
                 </div>
@@ -358,7 +358,7 @@ const UkToFrance: FC = () => {
                 <div className="tile-icon mx-auto">
                   <StaticImage
                     className="icon icon--responsive"
-                    src="../../images/routes/rwc-logo.256.jpg"
+                    src="../../images/regular-routes/rwc-logo.256.jpg"
                     alt="Frontline Group Logo: Refugee Women's Center (RWC)"
                   />
                 </div>
@@ -371,7 +371,7 @@ const UkToFrance: FC = () => {
                 <div className="tile-icon mx-auto">
                   <StaticImage
                     className="icon icon--responsive"
-                    src="../../images/routes/auberge-logo.256.jpg"
+                    src="../../images/regular-routes/auberge-logo.256.jpg"
                     alt="Frontline Group Logo: L'Auberge des Migrants (Auberge)"
                   />
                 </div>
@@ -432,7 +432,7 @@ const UkToFrance: FC = () => {
               <div className="mb-12 flex">
                 <div className="tile-icon">
                   <StaticImage
-                    src="../../images/routes/pallet-aid-logo.256.png"
+                    src="../../images/regular-routes/pallet-aid-logo.256.png"
                     alt="Hub Logo: Pallet Aid (PA)"
                     height={80}
                     width={80}
