+++ src/stylesheets/routes.css
@@ -1,7 +1,7 @@
 /*
 Legacy CSS for the routes pages
-/routes/uk-to-lebanon
-/routes/uk-to-france
+/regular-routes/uk-to-lebanon
+/regular-routes/uk-to-france
 */
 
 .section--content-left,
